function/requirements.txt
google-cloud-bigquery==3.3.2
colorama==0.4.5
pygments-pprint-sql==0.1.0
google-cloud-storage==2.5.0